+++ src/wasm.h
@@ -1055,7 +1055,6 @@ public:
   AtomicRMWOp op;
   uint8_t bytes;
   Address offset;
-  Address align;
   Expression* ptr;
   Expression* value;
   Name memory;
@@ -1071,7 +1070,6 @@ public:
 
   uint8_t bytes;
   Address offset;
-  Address align;
   Expression* ptr;
   Expression* expected;
   Expression* replacement;
@@ -1087,7 +1085,6 @@ public:
   AtomicWait(MixedArena& allocator) : AtomicWait() {}
 
   Address offset;
-  Address align;
   Expression* ptr;
   Expression* expected;
   Expression* timeout;
@@ -1103,7 +1100,6 @@ public:
   AtomicNotify(MixedArena& allocator) : AtomicNotify() {}
 
   Address offset;
-  Address align;
   Expression* ptr;
   Expression* notifyCount;
   Name memory;