+++ src/wasm/wasm-stack.cpp
@@ -548,7 +548,7 @@ void BinaryInstWriter::visitAtomicRMW(AtomicRMW* curr) {
     default:
       WASM_UNREACHABLE("unexpected op");
   }
-  emitMemoryAccess(curr->align,
+  emitMemoryAccess(curr->bytes,
                    curr->bytes,
                    curr->offset,
                    curr->memory,
@@ -595,7 +595,7 @@ void BinaryInstWriter::visitAtomicCmpxchg(AtomicCmpxchg* curr) {
     default:
       WASM_UNREACHABLE("unexpected type");
   }
-  emitMemoryAccess(curr->align,
+  emitMemoryAccess(curr->bytes,
                    curr->bytes,
                    curr->offset,
                    curr->memory,
@@ -608,7 +608,7 @@ void BinaryInstWriter::visitAtomicWait(AtomicWait* curr) {
   switch (curr->expectedType.getBasic()) {
     case Type::i32: {
       o << static_cast<int8_t>(BinaryConsts::I32AtomicWait);
-      emitMemoryAccess(curr->align,
+      emitMemoryAccess(4,
                        4,
                        curr->offset,
                        curr->memory,
@@ -618,7 +618,7 @@ void BinaryInstWriter::visitAtomicWait(AtomicWait* curr) {
     }
     case Type::i64: {
       o << static_cast<int8_t>(BinaryConsts::I64AtomicWait);
-      emitMemoryAccess(curr->align,
+      emitMemoryAccess(8,
                        8,
                        curr->offset,
                        curr->memory,
@@ -634,7 +634,7 @@ void BinaryInstWriter::visitAtomicWait(AtomicWait* curr) {
 void BinaryInstWriter::visitAtomicNotify(AtomicNotify* curr) {
   o << static_cast<int8_t>(BinaryConsts::AtomicPrefix)
     << static_cast<int8_t>(BinaryConsts::AtomicNotify);
-  emitMemoryAccess(curr->align,
+  emitMemoryAccess(4,
                    4,
                    curr->offset,
                    curr->memory,